+++ tests/SchemaTest.php
@@ -7,15 +7,13 @@ namespace Yiisoft\Db\Mysql\Tests;
 use Throwable;
 use Yiisoft\Db\Command\CommandInterface;
 use Yiisoft\Db\Connection\ConnectionInterface;
-use Yiisoft\Db\Constant\ColumnType;
-use Yiisoft\Db\Constant\PseudoType;
 use Yiisoft\Db\Constraint\Constraint;
 use Yiisoft\Db\Driver\Pdo\PdoConnectionInterface;
 use Yiisoft\Db\Exception\Exception;
 use Yiisoft\Db\Exception\InvalidConfigException;
 use Yiisoft\Db\Exception\NotSupportedException;
 use Yiisoft\Db\Expression\Expression;
-use Yiisoft\Db\Mysql\Column;
+use Yiisoft\Db\Mysql\Column\ColumnBuilder;
 use Yiisoft\Db\Mysql\Column\ColumnFactory;
 use Yiisoft\Db\Mysql\Schema;
 use Yiisoft\Db\Mysql\Tests\Support\TestTrait;
@@ -40,7 +38,7 @@ final class SchemaTest extends CommonSchemaTest
      *
      * @throws Exception
      */
-    public function testColumnSchema(array $columns, string $tableName): void
+    public function testColumns(array $columns, string $tableName): void
     {
         $db = $this->getConnection();
         $serverVersion = $db->getServerInfo()->getVersion();
@@ -76,15 +74,15 @@ final class SchemaTest extends CommonSchemaTest
             }
         }
 
-        parent::testColumnSchema($columns, $tableName);
+        parent::testColumns($columns, $tableName);
     }
 
     /**
      * @dataProvider \Yiisoft\Db\Mysql\Tests\Provider\SchemaProvider::columnsTypeBit
      */
-    public function testColumnSchemaWithTypeBit(array $columns): void
+    public function testColumnWithTypeBit(array $columns): void
     {
-        $this->columnSchema($columns, 'type_bit');
+        $this->assertTableColumns($columns, 'type_bit');
     }
 
     /**
@@ -104,19 +102,19 @@ final class SchemaTest extends CommonSchemaTest
         );
 
         $columnsData = [
-            'id' => ['int(11) NOT NULL AUTO_INCREMENT PRIMARY KEY', '', false],
-            'd' => ['date DEFAULT \'2011-11-11\'', '2011-11-11', false],
-            'dt' => ['datetime NOT NULL DEFAULT CURRENT_TIMESTAMP', 'CURRENT_TIMESTAMP', true],
-            'dt1' => ['datetime DEFAULT \'2011-11-11 00:00:00\'', '2011-11-11 00:00:00', false],
-            'dt2' => ['datetime DEFAULT CURRENT_TIMESTAMP', 'CURRENT_TIMESTAMP', true],
-            'ts' => ['timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP', 'CURRENT_TIMESTAMP', true],
-            'ts1' => ['timestamp DEFAULT \'2011-11-11 00:00:00\'', '2011-11-11 00:00:00', false],
-            'ts2' => ['timestamp DEFAULT CURRENT_TIMESTAMP', 'CURRENT_TIMESTAMP', true],
-            'simple_col' => ['varchar(40) DEFAULT \'uuid()\'', 'uuid()', false],
+            'id' => ['int(11) NOT NULL AUTO_INCREMENT PRIMARY KEY', ''],
+            'd' => ['date DEFAULT \'2011-11-11\'', '2011-11-11'],
+            'dt' => ['datetime NOT NULL DEFAULT CURRENT_TIMESTAMP', new Expression('CURRENT_TIMESTAMP')],
+            'dt1' => ['datetime DEFAULT \'2011-11-11 00:00:00\'', '2011-11-11 00:00:00'],
+            'dt2' => ['datetime DEFAULT CURRENT_TIMESTAMP', new Expression('CURRENT_TIMESTAMP')],
+            'ts' => ['timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP', new Expression('CURRENT_TIMESTAMP')],
+            'ts1' => ['timestamp DEFAULT \'2011-11-11 00:00:00\'', '2011-11-11 00:00:00'],
+            'ts2' => ['timestamp DEFAULT CURRENT_TIMESTAMP', new Expression('CURRENT_TIMESTAMP')],
+            'simple_col' => ['varchar(40) DEFAULT \'uuid()\'', 'uuid()'],
         ];
         if (!$oldMySQL) {
-            $columnsData['ts4'] = ['date DEFAULT (CURRENT_DATE + INTERVAL 2 YEAR)', '(curdate() + interval 2 year)', true];
-            $columnsData['uuid_col'] = ['varchar(40) DEFAULT (uuid())', 'uuid()', true];
+            $columnsData['ts4'] = ['date DEFAULT (CURRENT_DATE + INTERVAL 2 YEAR)', new Expression('(curdate() + interval 2 year)')];
+            $columnsData['uuid_col'] = ['varchar(40) DEFAULT (uuid())', new Expression('(uuid())')];
         }
 
         $columns = [];
@@ -135,12 +133,7 @@ final class SchemaTest extends CommonSchemaTest
 
         foreach ($tableSchema->getColumns() as $column) {
             $columnName = $column->getName();
-            if ($columnsData[$columnName][2]) {
-                $this->assertInstanceOf(Expression::class, $column->getDefaultValue());
-            } else {
-                $this->assertNotInstanceOf(Expression::class, $column->getDefaultValue());
-            }
-            $this->assertEquals($columnsData[$columnName][1], (string) $column->getDefaultValue());
+            $this->assertEquals($columnsData[$columnName][1], $column->getDefaultValue());
         }
     }
 
@@ -438,9 +431,9 @@ final class SchemaTest extends CommonSchemaTest
         $db->createCommand()->createTable(
             $tableName,
             [
-                'id' => new Column(PseudoType::PK),
-                'bool_col' => new Column(ColumnType::BOOLEAN),
-                'status' => new Column(ColumnType::TINYINT, 1),
+                'id' => ColumnBuilder::primaryKey(),
+                'bool_col' => ColumnBuilder::boolean(),
+                'status' => ColumnBuilder::tinyint(),
             ]
         )->execute();