+++ CHANGELOG.md
@@ -20,6 +20,11 @@
 - Enh #363: Refactor `Schema::normalizeDefaultValue()` method and move it to `ColumnFactory` class (@Tigrov)
 - Enh #366: Refactor `Quoter::quoteValue()` method (@Tigrov)
 - Chg #368: Update `QueryBuilder` constructor (@Tigrov)
+- Enh #367: Use `ColumnDefinitionBuilder` to generate table column SQL representation (@Tigrov)
+- Enh #371: Remove `ColumnInterface` (@Tigrov)
+- Enh #372: Rename `ColumnSchemaInterface` to `ColumnInterface` (@Tigrov)
+- Enh #373: Replace `DbArrayHelper::getColumn()` with `array_column()` (@Tigrov)
+- New #374: Add `IndexType` and `IndexMethod` classes (@Tigrov)
 
 ## 1.2.0 March 21, 2024